Release-1.7.6.1-20120122

+++ qtools/qcstring.h
@@ -87,10 +87,10 @@ Q_EXPORT void *qmemmove( void *dst, const void *src, uint len );
 Q_EXPORT char *qstrdup( const char * );
 
 Q_EXPORT inline uint cstrlen( const char *str )
-{ return strlen(str); }
+{ return (uint)strlen(str); }
 
 Q_EXPORT inline uint qstrlen( const char *str )
-{ return str ? strlen(str) : 0; }
+{ return str ? (uint)strlen(str) : 0; }
 
 Q_EXPORT inline char *cstrcpy( char *dst, const char *src )
 { return strcpy(dst,src); }
